deprecated mysql replaced with mysqli

+++ templates/delexercise.php
@@ -1,36 +1,42 @@
 <?php
 session_start();
 include_once("connection.php");
-$deleteID = $_POST["exerciseid"];
+include_once("login_required.php");
+
+$deleteID = $_POST["removeid"];
 
 $sql= /** @lang MYSQL */
     "select personid_fk,exerciseid
     from exercise
     where exerciseid = '$deleteID'";
 
-$retval = mysql_query($sql);
-$values = mysql_fetch_array($retval);
-$personid_fk = $values["personid_fk"];
 
-if($personid_fk != $_SESSION["personid"]){
-    echo "<h2>You trying to hack my shit??</h2>";
-    header("refresh:5;url=http://www.theworldsworstwebsiteever.com/" );
+$result = $conn->query($sql);
 
-}
-else{
-    $sql = /** @lang MYSQL */
-        "DELETE FROM exercise
-    WHERE exerciseid = '$deleteID'";
+if($result->num_rows > 0) {
+    while ($row = $result->fetch_assoc()) {
+        $personid_fk = $row["personid_fk"];
+        if($personid_fk != $_SESSION["personid"]){
+            echo "<h2>You trying to hack my shit??</h2>";
+            header("refresh:5;url=http://www.theworldsworstwebsiteever.com/" );
 
-    $retval = mysql_query($sql);
-    echo $retval;
-    if(! $retval) {
-        echo "<h2 style='text-align: center'>The exercise you are trying to delete is in use</h2>";
-        header("refresh:2;url=exercise.php" );
+        }
+        else{
+            $sql = /** @lang MYSQL */
+                "DELETE FROM exercise
+                WHERE exerciseid = '$deleteID'";
+
+
+            if($conn->query($sql) == TRUE){
+                echo $_POST["removeid"];
+
+            }
+            else{
+                echo "<h2 style='text-align: center'>The exercise you are trying to delete is in use</h2>";
+                header("refresh:2;url=exercise.php" );
+            }
+
+        }
     }
-    else{
-        echo "Entered data successfully\n";
-        header("Location: exercise.php");
-    }
-
 }
+
